move 'PastAppointmentList' directory to 'pages' directory
