+++ src/applications/vaos/appointment-list/pages/PastAppointmentsPage/PastAppointmentsDateDropdown.jsx
@@ -1,6 +1,6 @@
-import React, { useState } from 'react';
-import PropTypes from 'prop-types';
 import classNames from 'classnames';
+import PropTypes from 'prop-types';
+import React, { useState } from 'react';
 import Select from '../../../components/Select';
 
 function handleChange({ updateDateRangeIndex, callback }) {
