+++ src/applications/vaos/appointment-list/pages/PastAppointmentsPage/PastAppointmentsDateDropdown.unit.spec.js
@@ -1,10 +1,10 @@
-import React from 'react';
 import { expect } from 'chai';
 import moment from 'moment';
+import React from 'react';
 import sinon from 'sinon';
-import PastAppointmentsDateDropdown from './PastAppointmentsDateDropdown';
 import { getPastAppointmentDateRangeOptions } from '.';
 import { renderWithStoreAndRouter } from '../../../tests/mocks/setup';
+import PastAppointmentsDateDropdown from './PastAppointmentsDateDropdown';
 
 const ranges = getPastAppointmentDateRangeOptions(moment('2020-02-02'));
